Feature/gaggimate profile editor (#996)
* go: change ProfileSummary.ID to string, add Utility field

* go: update Adapter interface fakes for string profile IDs

* go: pass raw profile body through for GaggiMate create/update

* go: parse GaggiMate profile list/load/save with string ids and label field

* go: wire GaggiMate profile CRUD over live WS client

* go: enable ProfileEdit capability for GaggiMate profile handlers

* go: adjust Gaggiuino adapter for string profile ids

* frontend: add GaggiMate profile editor view and modal markup

* frontend: add GaggiMate phase overlay to shared chart plugin

* frontend: overlay GaggiMate named phases on shot history chart

* frontend: route profile actions to GaggiMate editor

* frontend: add GaggiMate editor and utility badge styles

* frontend: show utility badge and stale state in profile library

* i18n: add GaggiMate profile editor strings

* go: fix tests for GaggiMate ProfileEdit capability and string profile ids

* docs: update GaggiMate profile editing capability docs

* fix(codereview): patch XSS, listener leak, and fetch retry errors

* fix(gaggimate): patch XSS, listener leak, fetch retry, and CodeQL findings

- Escape p.id in data-id attributes (library-profile-editor) to prevent
  XSS from malicious GaggiMate profile IDs served over WebSocket
- Guard _bindInputs() body listener with _inputsBound flag — #gmEditorBody
  persists across renders so each _render() call was stacking handlers
- Wrap apiFetch in try-catch inside _fetchWithRetry loop; rethrows on
  final attempt or aborted signal, retries on transient network errors
- wire.go: use byte(v&0x7f) to make 7-bit varint mask explicit (CodeQL)
- ws.go: change wsGetProfileByID/wsDeleteProfile to accept uint32 directly;
  callers use ParseUint(..., 10, 32) — eliminates unchecked int->uint32
  narrowing flagged as high severity by CodeQL

* fix(gaggimate): address review findings from second pass

- gaggimate_live.go: reader goroutine now selects on ctx.Done() when
  sending to readCh so it can't block after the select loop exits;
  readErrCh send is non-blocking for the same reason — fixes goroutine
  and socket leak on every reconnect
- gaggiuino_adapter.go: propagate ParseUint error instead of silently
  using id 0, which caused the WS fallback to fetch the wrong profile
- handlers_profiles.go + validation.go: validate GaggiMate profile body
  (label non-empty, phases non-empty) before forwarding to device, same
  gate Gaggiuino already had via ProfileInput.Validate()
- handlers_profiles.go: PUT injects path {id} into body when absent so
  GaggiMate save updates rather than creating a duplicate
- gaggimate-profile-editor.js: clamp chart preview duration to 300s and
  add max attribute to prevent multi-million-point array on large values
- gaggimate-profile-editor.js: remove dead mode !== 'hold-*' conditionals
  inside block already gated on mode === 'pressure' || mode === 'flow'
- gaggimate_live.go: add TODO for rid-less dispatch FIFO correlation

+++ gaggiuino-local-profiler/go/internal/machines/ws.go
@@ -238,8 +238,8 @@ func wsGetProfileDict(ctx context.Context, baseURL string) ([]proto.SavedProfile
 	return dict.Profiles, nil
 }
 
-func wsGetProfileByID(ctx context.Context, baseURL string, id int) (*proto.ProfileDto, error) {
-	req := &proto.WebSocketProfileIdCommandDto{ID: uint32(id)}
+func wsGetProfileByID(ctx context.Context, baseURL string, id uint32) (*proto.ProfileDto, error) {
+	req := &proto.WebSocketProfileIdCommandDto{ID: id}
 	var profile proto.ProfileDto
 	if err := wsSendAndWait(ctx, baseURL, actionGetProfileByID, req.Marshal(), profile.Unmarshal); err != nil {
 		return nil, err
@@ -278,14 +278,14 @@ func wsUpdateProfile(ctx context.Context, baseURL string, profile ProfileInput) 
 	return proto.SavedProfileDto{}, fmt.Errorf("machine did not confirm profile id %d after update", *profile.ID)
 }
 
-func wsDeleteProfile(ctx context.Context, baseURL string, id int) ([]proto.SavedProfileDto, error) {
-	req := &proto.WebSocketProfileIdCommandDto{ID: uint32(id)}
+func wsDeleteProfile(ctx context.Context, baseURL string, id uint32) ([]proto.SavedProfileDto, error) {
+	req := &proto.WebSocketProfileIdCommandDto{ID: id}
 	var dict proto.SavedProfilesDto
 	if err := wsSendAndWait(ctx, baseURL, actionDeleteProfile, req.Marshal(), dict.Unmarshal); err != nil {
 		return nil, err
 	}
 	for _, p := range dict.Profiles {
-		if int(p.ID) == id {
+		if p.ID == id {
 			return nil, fmt.Errorf("machine did not confirm deletion of profile id %d", id)
 		}
 	}
